Refactored the way components are bound to a canvas.
Each component must be bound to canvas. Several canvas elements are
allowed, hence layering is now available. Several components may render
to a same layer, in which case the coordination must be ensured by the
components themselves.

If the canvas element does not exit, it will be created on the fly and
attached to the DOM. One may passa jQuery reference to an existing
canvas element, though.